requirements.txt
pillow
numpy
scikit-image
colorthief